Refactor: move shaped_abstractify to core
