+++ jax/_src/earray.py
@@ -16,7 +16,6 @@ from __future__ import annotations
 
 import math
 
-from jax._src import api_util
 from jax._src import basearray
 from jax._src import core
 from jax._src import tree_util
@@ -116,7 +115,7 @@ def _earray_shard_arg_handler(xs, shardings, layouts, copy_semantics):
   return pxla.shard_args(phys_shardings, layouts, copy_semantics, arrs)
 pxla.shard_arg_handlers[EArray] = _earray_shard_arg_handler
 
-api_util._shaped_abstractify_handlers[EArray] = lambda self: self.aval
+core.shaped_abstractify_handlers[EArray] = lambda self: self.aval
 core.pytype_aval_mappings[EArray] = lambda x: x.aval
 xla.canonicalize_dtype_handlers[EArray] = lambda x: x
 tree_util.dispatch_registry.register_node(
